requirements.txt
browser-use
langchain-openai
asyncio
python-dotenv
playwright